refactor: refactoring pipeline
