+++ core/elastic/domain.go
@@ -488,9 +488,9 @@ type ElasticsearchConfig struct {
 
 	AllowAccessWhenMasterNotFound bool `json:"allow_access_when_master_not_found,omitempty" config:"allow_access_when_master_not_found"`
 
-	BasicAuth *model.BasicAuth  `config:"basic_auth" json:"basic_auth,omitempty" elastic_mapping:"basic_auth:{type:object}"`
+	BasicAuth *model.BasicAuth `config:"basic_auth" json:"basic_auth,omitempty" elastic_mapping:"basic_auth:{type:object}"`
 	// Access token. Easysearch only.
-	Token     ucfg.SecretString `config:"token" json:"token,omitempty" yaml:"token" elastic_mapping:"token:{type:keyword}"`
+	Token ucfg.SecretString `config:"token" json:"token,omitempty" yaml:"token" elastic_mapping:"token:{type:keyword}"`
 
 	TrafficControl *struct {
 		Enabled              bool `json:"enabled,omitempty" config:"enabled"`
