+++ modules/pipeline/pipeline.go
@@ -1,580 +1,153 @@
-// Copyright (C) INFINI Labs & INFINI LIMITED.
-//
-// The INFINI Framework is offered under the GNU Affero General Public License v3.0
-// and as commercial software.
-//
-// For commercial licensing, contact us at:
-//   - Website: infinilabs.com
-//   - Email: hello@infini.ltd
-//
-// Open Source licensed under AGPL V3:
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <http://www.gnu.org/licenses/>.
+/* Copyright © INFINI LTD. All rights reserved.
+ * Web: https://infinilabs.com
+ * Email: hello#infini.ltd */
 
 package pipeline
 
 import (
-	"context"
-	"fmt"
-	"github.com/fsnotify/fsnotify"
-	"infini.sh/framework/core/locker"
-	"infini.sh/framework/core/task"
-	"runtime"
-	"sync"
-	"sync/atomic"
+	"net/http"
 	"time"
 
-	log "github.com/cihub/seelog"
-	"infini.sh/framework/core/api"
-	"infini.sh/framework/core/config"
-	"infini.sh/framework/core/env"
-	"infini.sh/framework/core/errors"
-	"infini.sh/framework/core/global"
+	httprouter "infini.sh/framework/core/api/router"
+	"infini.sh/framework/core/elastic"
+	"infini.sh/framework/core/orm"
 	"infini.sh/framework/core/pipeline"
-	"infini.sh/framework/core/rate"
-	"infini.sh/framework/core/util"
 )
 
-type PipeModule struct {
-	api.Handler
-	closed atomic.Bool
-
-	pipelines sync.Map
-	configs   sync.Map
-	contexts  sync.Map
-}
-
-func (module *PipeModule) Name() string {
-	return "pipeline"
-}
-
-var moduleCfg = struct {
-	PipelineEnabledByDefault bool `config:"pipeline_enabled_by_default"`
-}{PipelineEnabledByDefault: true}
-
-func (module *PipeModule) Setup() {
-	if global.Env().IsDebug {
-		log.Debug("pipeline framework config: ", moduleCfg)
-	}
-
-	ok, err := env.ParseConfig("preference", &moduleCfg)
-	if ok && err != nil && global.Env().SystemConfig.Configs.PanicOnConfigError {
-		panic(err)
-	}
-
-	module.pipelines = sync.Map{}
-	module.contexts = sync.Map{}
-	module.configs = sync.Map{}
-
-	pipeline.RegisterProcessorPlugin("dag", pipeline.NewDAGProcessor)
-	pipeline.RegisterProcessorPlugin("echo", NewEchoProcessor)
-
-	api.HandleAPIMethod(api.GET, "/pipeline/tasks/", module.getPipelinesHandler)
-	api.HandleAPIMethod(api.POST, "/pipeline/tasks/_search", module.searchPipelinesHandler)
-	api.HandleAPIMethod(api.POST, "/pipeline/tasks/", module.createPipelineHandler)
-	api.HandleAPIMethod(api.GET, "/pipeline/task/:id", module.getPipelineHandler)
-	api.HandleAPIMethod(api.DELETE, "/pipeline/task/:id", module.deletePipelineHandler)
-	api.HandleAPIMethod(api.POST, "/pipeline/task/:id/_start", module.startTaskHandler)
-	api.HandleAPIMethod(api.POST, "/pipeline/task/:id/_stop", module.stopTaskHandler)
-
-}
-
-func (module *PipeModule) startTask(taskID string) (exists bool) {
-	if module.closed.Load() {
-		return false
-	}
-
-	ctx, ok := module.contexts.Load(taskID)
-	if !ok {
-		return
-	}
-	v1, ok := ctx.(*pipeline.Context)
-	if !ok {
-		return
-	}
-
-	exists = true
-
-	// Mark exited pipeline to start again
-	if v1.IsExit() {
-		v1.Restart()
-	}
-	// Resume pipeline loop
-	if v1.IsPause() {
-		// Mark pipeline status as starting
-		v1.Starting()
-		v1.Resume()
-	}
-
-	return
-}
-
-// stopTask will cancel the current pipeline context, abort the pipeline execution.
-func (module *PipeModule) stopTask(taskID string) (exists bool) {
-	ctx, ok := module.contexts.Load(taskID)
-	if !ok {
-		return
-	}
-	v1, ok := ctx.(*pipeline.Context)
-	if !ok {
-		return
-	}
-
-	exists = true
-
-	if global.Env().IsDebug {
-		if rate.GetRateLimiterPerSecond("pipeline", "shutdown "+taskID+string(v1.GetRunningState()), 1).Allow() {
-			log.Trace("start shutting down pipeline:", taskID, ",state:", v1.GetRunningState())
-		}
-	}
-
-	// Mark pipeline as exited
-	v1.Exit()
-	// Mark pipeline as STOPPING as needed
-	v1.Stopping()
-	// call cancelFunc(), will mark IsCanceled asynchronously
-	v1.CancelTask()
-
-	return
-}
-
-// deleteTask will clean all in-memory states and release the pipeline context
-func (module *PipeModule) deleteTask(taskID string) {
-	module.pipelines.Delete(taskID)
-	module.configs.Delete(taskID)
-	module.releaseContext(taskID)
-	module.contexts.Delete(taskID)
-}
-
-// releaseContext will release the task context
-func (module *PipeModule) releaseContext(taskID string) {
-	ctx, ok := module.contexts.Load(taskID)
-	if ok {
-		v1, ok := ctx.(*pipeline.Context)
-		if ok {
-			pipeline.ReleaseContext(v1)
-			if v1.IsPause() {
-				// release loop
-				v1.Resume()
-			}
-		}
-	}
-}
-
-func getPipelineConfig() ([]pipeline.PipelineConfigV2, error) {
-	configFile := global.Env().GetConfigFile()
-	configDir := global.Env().GetConfigDir()
-	parentCfg, err := config.LoadFile(configFile)
+func (h *PipeModule) createPipelineHandler(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
+	var obj = &pipeline.PipelineConfigV2{}
+	err := h.DecodeJSON(req, obj)
 	if err != nil {
-		return nil, fmt.Errorf("failed to load config file: %v, path: %s", err, configFile)
-	}
-	childCfg, err := config.LoadPath(configDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load config dir: %v, path: %s", err, configDir)
-	}
-	err = parentCfg.Merge(childCfg)
-	if err != nil {
-		return nil, fmt.Errorf("failed to merge configs: %v", err)
+		h.WriteError(w, err.Error(), http.StatusInternalServerError)
+		return
 	}
 
-	pipelineCfg := []pipeline.PipelineConfigV2{}
+	ctx := orm.NewContextWithParent(req.Context())
+	ctx.Refresh = orm.WaitForRefresh
 
-	if ok := parentCfg.HasField("pipeline"); ok {
-		parentCfg, err = parentCfg.Child("pipeline", -1)
-		if err != nil {
-			return nil, err
-		}
-		err = parentCfg.Unpack(&pipelineCfg)
-		return pipelineCfg, err
+	err = orm.Create(ctx, obj)
+	if err != nil {
+		h.WriteError(w, err.Error(), http.StatusInternalServerError)
+		return
 	}
-	return pipelineCfg, nil
+
+	h.WriteCreatedOKJSON(w, obj.ID)
 }
 
-func (module *PipeModule) Start() error {
-	var (
-		pipelines []pipeline.PipelineConfigV2
-		err       error
-	)
-	pipelines, err = getPipelineConfig()
-	if err != nil && global.Env().SystemConfig.Configs.PanicOnConfigError {
-		panic(err)
-	}
-	for _, v := range pipelines {
-		err := module.createPipeline(v, false)
-		if err != nil {
-			log.Errorf("error on running pipeline: %v, err: %v", v.Name, err)
-			continue
-		}
+func (h *PipeModule) getPipelineHandler(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
+	id := ps.MustGetParameter("id")
+
+	obj := pipeline.PipelineConfigV2{}
+	obj.ID = id
+
+	ctx := orm.NewContextWithParent(req.Context())
+	exists, err := orm.GetV2(ctx, &obj)
+	if !exists || err != nil {
+		h.WriteOpRecordNotFoundJSON(w, id)
+		return
 	}
 
-	//listen on changes
-	config.NotifyOnConfigChange(func(ev fsnotify.Event) {
-		if module.closed.Load() || global.ShuttingDown() {
-			log.Warn("module closed, skip reloading pipelines")
+	h.WriteOpRecordNotFoundJSON(w, id)
+}
+
+func (h *PipeModule) updatePipelineHandler(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
+	id := ps.MustGetParameter("id")
+	obj := pipeline.PipelineConfigV2{}
+
+	replace := h.GetBoolOrDefault(req, "replace", false)
+	ctx := orm.NewContextWithParent(req.Context())
+
+	var err error
+	var create *time.Time
+	if !replace {
+		obj.ID = id
+
+		//can't remove, since we need it for update
+		exists, err := orm.GetWithSystemFields(ctx, &obj)
+		if !exists || err != nil {
+			h.WriteOpRecordNotFoundJSON(w, id)
 			return
 		}
-
-		log.Infof("config changed, checking for new pipeline configs, %v, %v", ev.Op, ev.Name)
-
-		newConfig := []pipeline.PipelineConfigV2{}
-		newConfig, err = getPipelineConfig()
-
-		if err != nil {
-			log.Error(err)
-			return
-		}
-
-		defer func() {
-			if !global.Env().IsDebug {
-				if r := recover(); r != nil {
-					var v string
-					switch r.(type) {
-					case error:
-						v = r.(error).Error()
-					case runtime.Error:
-						v = r.(runtime.Error).Error()
-					case string:
-						v = r.(string)
-					}
-					log.Error("error on apply pipeline change,", v)
-				}
-			}
-		}()
-
-		needStopAndClean := []string{}
-		newPipelines := map[string]pipeline.PipelineConfigV2{}
-
-		newPipelineNames := []string{}
-		for _, v := range newConfig {
-			newPipelines[v.Name] = v
-			newPipelineNames = append(newPipelineNames, v.Name)
-		}
-
-		log.Debugf("we now have %v new pipelines: %v", len(newPipelineNames), newPipelineNames)
-
-		module.configs.Range(func(k, v any) bool {
-			oldC, ok := v.(pipeline.PipelineConfigV2)
-			if !ok {
-				log.Warnf("impossible value from configs: %v", v)
-				return true
-			}
-
-			// Don't stop transient pipelines
-			if oldC.Transient {
-				log.Debugf("transient pipeline %v should not be reloaded", oldC.Name)
-				return true
-			}
-			newC, ok := newPipelines[oldC.Name]
-			isSame := newC.Equals(oldC)
-			// Skip condition: (old pipeline is present in the new pipeline configs, config is the same, new config is also enabled)
-			if ok && isSame && isPipelineEnabled(newC.Enabled) {
-				log.Debugf("pipeline %v config not changed, skip reloading", oldC.Name)
-				return true
-			}
-
-			log.Debug("pipeline config changed, stop and clean:", oldC.Name, ",", oldC, ",", ok, ",", isSame, ",", isPipelineEnabled(newC.Enabled))
-
-			needStopAndClean = append(needStopAndClean, oldC.Name)
-			return true
-		})
-
-		log.Debugf("we now have %v old pipelines need to clean: %v", len(needStopAndClean), needStopAndClean)
-
-		if len(needStopAndClean) > 0 {
-			log.Trace("stop and wait for pipelines to release: ", needStopAndClean)
-			module.stopAndWaitForRelease(needStopAndClean, time.Minute)
-			log.Debug("old pipelines released")
-
-			for _, taskID := range needStopAndClean {
-				log.Infof("removing pipeline [%s]", taskID)
-				module.deleteTask(taskID)
-			}
-		}
-
-		log.Debugf("starting %v pipelines", len(newPipelineNames))
-		for k, v := range newPipelines {
-			err := module.createPipeline(v, false)
-			if err != nil {
-				log.Errorf("failed to create pipeline: %v, err: %v", k, err)
-			}
-		}
-	})
-
-	return nil
-}
-
-func (module *PipeModule) Stop() error {
-	if module.closed.Load() {
-		return nil
-	}
-	module.closed.Store(true)
-
-	total := util.GetSyncMapSize(&module.contexts)
-	if total <= 0 {
-		return nil
-	}
-
-	log.Info("shutting down pipelines")
-
-	var taskIDs []string
-	module.contexts.Range(func(key, value any) bool {
-		taskID, ok := key.(string)
-		if !ok {
-			return false
-		}
-		taskIDs = append(taskIDs, taskID)
-		return true
-	})
-
-	module.stopAndWaitForRelease(taskIDs, time.Minute*5)
-
-	log.Info("finished shut down pipelines")
-	return nil
-}
-
-func (module *PipeModule) stopAndWaitForRelease(taskIDs []string, timeout time.Duration) {
-	start := time.Now()
-
-	for {
-		if time.Now().Sub(start) > timeout {
-			log.Error("waitForStop timed out")
-			break
-		}
-
-		// Send stop signal to all contexts
-		for _, taskID := range taskIDs {
-			// cancel & stop
-			module.stopTask(taskID)
-			// release loop
-			module.releaseContext(taskID)
-			// don't delete context yet
-		}
-
-		needRetry := false
-		for _, taskID := range taskIDs {
-			v, ok := module.contexts.Load(taskID)
-			if !ok {
-				continue
-			}
-
-			ctx, ok := v.(*pipeline.Context)
-			if !ok {
-				log.Errorf("impossible value from contexts: %t", v)
-				continue
-			}
-
-			if !ctx.IsLoopReleased() {
-				if rate.GetRateLimiterPerSecond("pipeline", "shutdown"+taskID+string(ctx.GetRunningState()), 1).Allow() {
-					log.Debug("pipeline still running: ", taskID, ",state: ", ctx.GetRunningState())
-				}
-				needRetry = true
-				break
-			}
-		}
-
-		if !needRetry {
-			break
-		}
-	}
-}
-
-const pipelineSingleton = "pipeline_singleton"
-
-var creatingLocker = sync.Mutex{}
-
-func (module *PipeModule) createPipeline(v pipeline.PipelineConfigV2, transient bool) error {
-	if module.closed.Load() {
-		return errors.New("module closed")
-	}
-
-	if !isPipelineEnabled(v.Enabled) {
-		// pipeline config explicitly disabled
-		return nil
-	}
-
-	creatingLocker.Lock()
-	defer creatingLocker.Unlock()
-
-	v.Transient = transient
-
-	// NOTE: hold the slot before creating pipeline loops
-	if _, ok := module.configs.LoadOrStore(v.Name, v); ok {
-		log.Tracef("pipeline [%v] is already created, skip", v.Name)
-		return nil
-	}
-
-	if v.Singleton {
-		log.Info("creating pipeline: " + v.Name + ", singleton")
+		id = obj.ID
+		create = obj.Created
 	} else {
-		log.Info("creating pipeline: " + v.Name)
+		t := time.Now()
+		create = &t
 	}
 
-	task.RunWithContext("pipeline:"+v.Name, func(taskCtx context.Context) error {
+	obj = pipeline.PipelineConfigV2{}
+	err = h.DecodeJSON(req, &obj)
+	if err != nil {
+		h.WriteError(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
 
-		cfgV := taskCtx.Value("cfg")
-		cfg, ok := cfgV.(pipeline.PipelineConfigV2)
-		if !ok {
-			return errors.New("invalid pipeline config")
-		}
-		processors, err := cfg.GetProcessorsConfig()
-		if err != nil {
-			return errors.Errorf("failed to get processor config, %v", err)
-		}
+	//protect
+	obj.ID = id
+	obj.Created = create
 
-		processor, err := pipeline.NewPipeline(processors)
-		if err != nil {
-			return err
-		}
+	ctx.Refresh = orm.WaitForRefresh
+	err = orm.Save(ctx, &obj)
+	if err != nil {
+		h.WriteError(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
 
-		ctx := pipeline.AcquireContext(v)
-		module.pipelines.Store(v.Name, processor)
-		module.contexts.Store(v.Name, ctx)
-
-		defer func() {
-			if !global.Env().IsDebug {
-				if r := recover(); r != nil {
-					var err string
-					switch r.(type) {
-					case error:
-						err = r.(error).Error()
-					case runtime.Error:
-						err = r.(runtime.Error).Error()
-					case string:
-						err = r.(string)
-					}
-					log.Errorf("error on pipeline: %v, retry delay: %vms", cfg.Name, err)
-				}
-			}
-
-			ctx.SetLoopReleased()
-			processor.Release()
-		}()
-
-		if !cfg.AutoStart {
-			// Mark pipeline as exited, don't run automatically
-			ctx.Exit()
-		} else {
-			ctx.Starting()
-		}
-
-		log.Debug("processing pipeline_v2: ", cfg.Name)
-
-		retryDelayInMs := 1000
-		if cfg.RetryDelayInMs > 0 {
-			retryDelayInMs = cfg.RetryDelayInMs
-		}
-
-		started := false
-
-		for {
-			if global.ShuttingDown() {
-				log.Debugf("system is shutting down, pipeline [%v] will be stopped", cfg.Name)
-				break
-			}
-
-			if ctx.IsReleased() {
-				break
-			}
-			if module.closed.Load() {
-				break
-			}
-			if ctx.IsExit() {
-				ctx.Stopped()
-			}
-
-			// NOTE: state must be checked as the last step
-			state := ctx.GetRunningState()
-			if global.Env().IsDebug {
-				log.Tracef("pipeline [%v], state: %v", cfg.Name, state)
-				time.Sleep(500 * time.Millisecond)
-			}
-			switch state {
-			case pipeline.STARTING:
-
-				//check
-				if v.Singleton {
-					if v.MaxRunningInMs <= 0 {
-						v.MaxRunningInMs = 60000
-					}
-					ok, err := locker.Hold(pipelineSingleton, v.Name, global.Env().SystemConfig.NodeConfig.ID, time.Duration(v.MaxRunningInMs)*time.Millisecond, true)
-					if !ok {
-						log.Debugf("pipeline [%v] is already running somewhere, %v", cfg.Name, err)
-						ctx.Finished()
-						continue
-					}
-				}
-
-				// Pipeline needs to run
-				if started {
-					log.Errorf("pipeline [%v] started twice, should not happen", cfg.Name)
-				}
-				started = true
-				ctx.Started()
-				ctx.ResetContext()
-
-				err = processor.Process(ctx)
-
-				if err != nil {
-					log.Errorf("error on pipeline:%v, %v", cfg.Name, err)
-					ctx.Failed(err)
-				} else {
-					if global.Env().IsDebug {
-						log.Debugf("pipeline [%v] end running", cfg.Name)
-					}
-					ctx.Finished()
-				}
-				started = false
-			case pipeline.STARTED, pipeline.STOPPING:
-				log.Errorf("pipeline [%v] loop should not detect %s", cfg.Name, state)
-			case pipeline.FINISHED, pipeline.FAILED:
-				// Pipeline ended, pause or start next round
-				// keep_running: true & not stopped manually by Exit()
-				// For IsExit, don't pause here, wait for STOPPED state, or we could Pause twice for STOPPED & IsExit.
-				if cfg.KeepRunning {
-					if global.Env().IsDebug {
-						log.Tracef("pipeline [%v] end running, restart again, retry in [%v]ms", cfg.Name, retryDelayInMs)
-					}
-
-					if retryDelayInMs > 0 {
-						log.Tracef("start sleep [%v]ms", retryDelayInMs)
-						time.Sleep(time.Duration(retryDelayInMs) * time.Millisecond)
-						log.Tracef("end sleep [%v]ms", retryDelayInMs)
-					}
-
-					// restart after delay.
-					ctx.Starting()
-				} else {
-					ctx.Stopped()
-					ctx.Pause()
-				}
-			case pipeline.STOPPED:
-				// Pipeline manually stopped, pause
-				ctx.Pause()
-			}
-		}
-
-		log.Debugf("pipeline [%v] loop exited with state [%v]", cfg.Name, ctx.GetRunningState())
-
-		return nil
-	}, context.WithValue(context.Background(), "cfg", v))
-
-	return nil
+	h.WriteUpdatedOKJSON(w, obj.ID)
 }
 
-func isPipelineEnabled(enabled *bool) bool {
-	// if not configured `enabled: true`, by default true
-	if enabled == nil {
-		return moduleCfg.PipelineEnabledByDefault
+func (h *PipeModule) deletePipelineHandler(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
+	id := ps.MustGetParameter("id")
+
+	obj := pipeline.PipelineConfigV2{}
+	obj.ID = id
+
+	ctx := orm.NewContextWithParent(req.Context())
+
+	exists, err := orm.GetV2(ctx, &obj)
+	if !exists || err != nil {
+		h.WriteOpRecordNotFoundJSON(w, id)
+		return
 	}
-	return *enabled
+
+	ctx.Refresh = orm.WaitForRefresh
+	err = orm.Delete(ctx, &obj)
+	if err != nil {
+		h.WriteError(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	h.WriteDeletedOKJSON(w, id)
+}
+
+func (h *PipeModule) searchPipelineHandler(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
+
+	var err error
+	builder, err := orm.NewQueryBuilderFromRequest(req, "name", "combined_fulltext")
+	if err != nil {
+		h.WriteError(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	if len(builder.Sorts()) == 0 {
+		builder.SortBy(orm.Sort{Field: "created", SortType: orm.DESC})
+	}
+
+	builder.EnableBodyBytes()
+
+	ctx := orm.NewContextWithParent(req.Context())
+	orm.WithModel(ctx, &pipeline.PipelineConfigV2{})
+
+	var objs []pipeline.PipelineConfigV2
+
+	err, res := elastic.SearchV2WithResultItemMapper(ctx, &objs, builder, nil)
+	if err != nil {
+		h.WriteError(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	_, err = h.Write(w, res.Raw)
+	if err != nil {
+		h.Error(w, err)
+	}
 }
